+++ src/Test/StatementRepository.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the xAPI package.
+ *
+ * (c) Christian Flothmann <christian.flothmann@xabbuh.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace XApi\Repository\Doctrine\Test;
 
 use Doctrine\Common\Persistence\ObjectManager;
@@ -25,6 +34,9 @@ final class StatementRepository implements StatementRepositoryInterface
         $this->objectManager = $objectManager;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function findStatementById(StatementId $statementId, Actor $authority = null)
     {
         $statement = $this->repository->findStatementById($statementId, $authority);
@@ -33,25 +45,36 @@ final class StatementRepository implements StatementRepositoryInterface
         return $statement;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function findVoidedStatementById(StatementId $voidedStatementId, Actor $authority = null)
     {
-        $statement = $this->findVoidedStatementById($voidedStatementId, $authority);
+        $statement = $this->repository->findVoidedStatementById($voidedStatementId, $authority);
         $this->objectManager->clear();
 
         return $statement;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function findStatementsBy(StatementsFilter $criteria, Actor $authority = null)
     {
-        $statements = $this->findStatementsBy($criteria, $authority);
+        $statements = $this->repository->findStatementsBy($criteria, $authority);
         $this->objectManager->clear();
 
         return $statements;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function storeStatement(Statement $statement, $flush = true)
     {
-        $this->repository->storeStatement($statement);
+        $statementId = $this->repository->storeStatement($statement);
         $this->objectManager->clear();
+
+        return $statementId;
     }
 }